+++ authidentity/src/main/java/aivle/domain/entity/AdminAccount.java
@@ -14,10 +14,10 @@ import aivle.AuthidentityApplication;
 import aivle.domain.command.LoginCommand;
 import aivle.domain.command.LogoutCommand;
 import aivle.domain.command.SignupCommand;
-import aivle.domain.enums.UserRole;
 import aivle.domain.event.Logged;
 import aivle.domain.event.SignedUp;
 import aivle.domain.repository.AdminAccountRepository;
+import aivle.domain.valueobject.UserRole;
 import lombok.Data;
 
 @Entity
