+++ authidentity/src/main/java/aivle/domain/valueobject/UserRole.java
@@ -1,4 +1,4 @@
-package aivle.domain.enums;
+package aivle.domain.valueobject;
 
 public enum UserRole {
     ADMIN, AUTHOR, USER
