+++ authidentity/src/main/java/aivle/infra/processor/AuthorAccountHateoasProcessor.java
@@ -1,10 +1,11 @@
-package aivle.infra;
+package aivle.infra.processor;
 
-import aivle.domain.*;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.Link;
 import org.springframework.hateoas.server.RepresentationModelProcessor;
 import org.springframework.stereotype.Component;
+
+import aivle.domain.entity.AuthorAccount;
 
 @Component
 public class AuthorAccountHateoasProcessor
